refactor(data-point/core): core/factory.js -> core/core.js
closes #145
