+++ packages/data-point/lib/core/index.js
@@ -1,5 +1,5 @@
 /* eslint global-require: 0 */
 
 module.exports = {
-  create: require('./factory').create
+  create: require('./core').create
 }
